backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
sqlalchemy==2.0.36
pymysql==1.1.1
redis==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
scikit-learn==1.5.2
numpy>=1.26.4
openai>=1.30.0
sentence-transformers>=3.0.0
slowapi>=0.1.9
PyPDF2>=3.0.0
python-docx>=1.1.0
bleach>=6.1.0
